+++ modules/light-clients/09-localhost/types/client_state.go
@@ -79,28 +79,28 @@ func (cs ClientState) ExportMetadata(_ sdk.KVStore) []exported.GenesisMetadata {
 
 // CheckHeaderAndUpdateState updates the localhost client. It only needs access to the context
 func (cs *ClientState) CheckHeaderAndUpdateState(
-	ctx sdk.Context, cdc codec.BinaryCodec, clientStore sdk.KVStore, header exported.Header,
+	ctx sdk.Context, cdc codec.BinaryCodec, clientStore sdk.KVStore, header exported.ClientMessage,
 ) (exported.ClientState, exported.ConsensusState, error) {
 	return cs.UpdateState(ctx, cdc, clientStore, header)
 }
 
 // VerifyHeader is a no-op.
 func (cs *ClientState) VerifyHeader(
-	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.Header,
+	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.ClientMessage,
 ) (exported.ClientState, exported.ConsensusState, error) {
 	return cs, nil, nil
 }
 
 // CheckForMisbehaviour returns false.
 func (cs *ClientState) CheckForMisbehaviour(
-	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.Header,
+	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.ClientMessage,
 ) (bool, error) {
 	return false, nil
 }
 
 // UpdateState updates the localhost client. It only needs access to the context
 func (cs *ClientState) UpdateState(
-	ctx sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.Header,
+	ctx sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.ClientMessage,
 ) (exported.ClientState, exported.ConsensusState, error) {
 	// use the chain ID from context since the localhost client is from the running chain (i.e self).
 	cs.ChainId = ctx.ChainID()
@@ -111,7 +111,7 @@ func (cs *ClientState) UpdateState(
 
 // UpdateStateOnMisbehaviour returns an error (no misbehaviour case).
 func (cs *ClientState) UpdateStateOnMisbehaviour(
-	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.Header,
+	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.ClientMessage,
 ) (*ClientState, error) {
 	return nil, sdkerrors.Wrapf(clienttypes.ErrUpdateClientFailed, "cannot update localhost client on misbehaviour")
 }
@@ -120,7 +120,7 @@ func (cs *ClientState) UpdateStateOnMisbehaviour(
 // Since localhost is the client of the running chain, misbehaviour cannot be submitted to it
 // Thus, CheckMisbehaviourAndUpdateState returns an error for localhost
 func (cs ClientState) CheckMisbehaviourAndUpdateState(
-	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.Misbehaviour,
+	_ sdk.Context, _ codec.BinaryCodec, _ sdk.KVStore, _ exported.ClientMessage,
 ) (exported.ClientState, error) {
 	return nil, sdkerrors.Wrap(clienttypes.ErrInvalidMisbehaviour, "cannot submit misbehaviour to localhost client")
 }